pyproject.toml
[project]
name = "trace"
version = "0.1.0"
description = "Record, synthesize, and replay on-screen macOS workflows."
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.8",
    "anthropic>=0.40",
    "jsonschema>=4.23",
    # Server-Sent Events transport for the synthesizer's question-and-answer
    # endpoints — see services/synthesizer/src/synthesizer/api.py.
    "sse-starlette>=2.1",
    "trace-recorder",
    "trace-synthesizer",
    "trace-runner",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "mypy>=1.11",
    "ruff>=0.6",
    "types-jsonschema>=4.23",
    # respx mocks httpx clients (used to stub the Anthropic SDK in runner & synth tests).
    "respx>=0.21",
    # pillow round-trips preview PNGs in snapshot/golden-fixture tests.
    "pillow>=10.4",
]

[tool.uv.workspace]
members = ["services/recorder", "services/synthesizer", "services/runner"]

[tool.uv.sources]
trace-recorder = { workspace = true }
trace-synthesizer = { workspace = true }
trace-runner = { workspace = true }

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = [".venv", "build", "dist"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.isort]
known-first-party = ["recorder", "synthesizer", "runner", "gateway"]

[tool.ruff.lint.per-file-ignores]
# draft_prompt.py / revise_prompt.py embed the full LLM system prompts as
# multi-line strings; prose paragraphs regularly exceed 100 chars and
# hard-wrapping them would only obscure what the model actually sees.
# similarity.py contains long descriptive failure messages for snapshot diffs.
"services/synthesizer/src/synthesizer/draft_prompt.py" = ["E501"]
"services/synthesizer/src/synthesizer/revise_prompt.py" = ["E501"]
"services/synthesizer/src/synthesizer/similarity.py" = ["E501"]
# Stage-1 merge: a number of runner test files now embed canonical
# notes_daily params (`{"note_template": "- [ ] focus block\n"}`) inline
# inside JSON request bodies that already had ``"skill_slug": "notes_daily"``.
# Wrapping each call site would obscure the test intent; allow long lines
# in the test layer.
"tests/runner/test_api.py" = ["E501"]
"tests/runner/test_e2e_fake.py" = ["E501"]
"tests/runner/test_budget_config.py" = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["services", "gateway"]
# All three modules now use a src/ layout for their packages.
# mypy_path lets the type checker resolve the in-tree packages.
mypy_path = "services/recorder/src:services/synthesizer/src:services/runner/src"
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "Cocoa.*",
    "Quartz.*",
    "ApplicationServices.*",
    "AppKit.*",
    "Foundation.*",
    "objc.*",
    "respx.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["services/recorder/src", "services/synthesizer/src", "services/runner/src"]
markers = [
    "macos: integration tests that require a real macOS host with Accessibility/Input Monitoring permissions granted; skipped by default off-darwin or without permissions.",
    "uses_lsappinfo: tests that exercise the ``lsappinfo`` frontmost-app helper directly and opt out of the module-level stub that otherwise forces it to return None.",
    "live_input: test posts real events via CGEventPost; skipped unless --run-live-input is passed.",
]
